+++ app/com/lynxanalytics/biggraph/graph_operations/EdgesFromAttributeMatches.scala
@@ -3,6 +3,7 @@ package com.lynxanalytics.biggraph.graph_operations
 
 import com.lynxanalytics.biggraph.graph_api._
 import com.lynxanalytics.biggraph.spark_util.Implicits._
+import com.lynxanalytics.biggraph.spark_util.HybridRDD
 import org.apache.spark.Partitioner
 
 object EdgesFromAttributeMatches extends OpFromJson {
@@ -98,7 +99,7 @@ object EdgesFromUniqueBipartiteAttributeMatches extends OpFromJson {
     val edges = edgeBundle(
       inputs.from.entity,
       inputs.to.entity,
-      EdgeBundleProperties.partialFunction)
+      EdgeBundleProperties.matching)
   }
   def fromJson(j: JsValue) =
     EdgesFromUniqueBipartiteAttributeMatches()
@@ -131,6 +132,60 @@ case class EdgesFromUniqueBipartiteAttributeMatches()
       .values
       .map { case (fromId, toId) => (fromId, Edge(fromId, toId)) }
       .sortUnique(partitioner)
+
+    output(o.edges, mapping)
+  }
+}
+
+// Generates edges between vertices that match on an attribute.
+// If fromAttr on A matches toAttr on B, an A -> B edge is generated.
+// The values of toAttr have to be unique, otherwise this operation fails.
+object EdgesFromLookupAttributeMatches extends OpFromJson {
+  class Input extends MagicInputSignature {
+    val from = vertexSet
+    val fromAttr = vertexAttribute[String](from)
+    val to = vertexSet
+    val toAttr = vertexAttribute[String](to)
+  }
+  class Output(implicit instance: MetaGraphOperationInstance,
+               inputs: Input)
+      extends MagicOutput(instance) {
+    val edges = edgeBundle(
+      inputs.from.entity,
+      inputs.to.entity,
+      EdgeBundleProperties.partialFunction)
+  }
+  def fromJson(j: JsValue) =
+    EdgesFromLookupAttributeMatches()
+}
+case class EdgesFromLookupAttributeMatches()
+    extends TypedMetaGraphOp[EdgesFromLookupAttributeMatches.Input, EdgesFromLookupAttributeMatches.Output] {
+  import EdgesFromLookupAttributeMatches._
+
+  override val isHeavy = true
+  @transient override lazy val inputs = new Input()
+  def outputMeta(instance: MetaGraphOperationInstance) =
+    new Output()(instance, inputs)
+
+  def execute(inputDatas: DataSet,
+              o: Output,
+              output: OutputBuilder,
+              rc: RuntimeContext): Unit = {
+    implicit val id = inputDatas
+    implicit val runtimeContext = rc
+    val partitioner = EdgesFromBipartiteAttributeMatches.getLargerPartitioner(
+      inputs.from.rdd.partitioner.get,
+      inputs.to.rdd.partitioner.get)
+    val fromStringToId = inputs.fromAttr.rdd
+      .map(_.swap)
+    val toStringToId = inputs.toAttr.rdd
+      .map(_.swap)
+      .assertUniqueKeys(partitioner)
+    val mapping = HybridRDD(fromStringToId, partitioner, even = true)
+      .lookup(toStringToId)
+      .values
+      .map { case (fromId, toId) => Edge(fromId, toId) }
+      .randomNumbered(partitioner)
 
     output(o.edges, mapping)
   }